refactor(imessage): move GUID normalization helper
